+++ docker-compose.yaml
@@ -1,0 +1,32 @@
+version: "3.8"
+services:
+   
+  backend:
+    build: ./backend
+    ports:
+      - "5000:5000"
+    volumes:
+      - logs:/app/logs
+      - ./backend:/app
+      - /app/node_modules
+   
+  frontend:
+    build: ./frontend
+    ports:
+      - "3000:3000"
+    volumes:
+      - ./frontend/src:/app/src
+    stdin_open: true
+    tty: true
+    depends_on:
+      - backend
+  mongo:
+      image: mongo
+      ports:
+        - "27018:27018"
+      command: mongod --port 27018 --bind_ip_all
+
+ 
+volumes:
+  data:
+  logs:
